Create new mojos to split up prepare, upload steps
Ran into an issue where a consequence of splitting up the package and
deploy steps resulted in the the files listed in the classpath field in
the manifest file not being the same ones as were used in the Slimfast
file (which ultimately became the files on disk). This manifested in
NoClassDefFound errors for files in certain jars.

This allows writing a manifest file to disk (intended to be during the
package phase) that will list exactly what was used during the build,
and then a second mojo (for the deploy phase) that will read from that
manifest file, then do the upload and write out the normal Slimfast
manifest file for later distribution.